# openapi.yaml
openapi: 3.1.0

info:
  title: Crowd Transcription Fixer API
  version: 1.0.0
  description: |
    REST API for a crowd-sourced subtitle correction service.

    **How it works**
    1. Audio metadata is synced from an external MSSQL database on startup.
    2. Clients browse available audios and claim one to work on — this creates a
       *task* that locks the audio for up to 2 hours.
    3. The client plays the audio via the URL stored from MSSQL and fetches the
       original VTT subtitle through the backend (served from S3, authorised by
       the session token).
    4. The client submits the corrected VTT; the backend writes it to S3 and
       marks the audio completed.

    **Authentication**
    Public endpoints (browsing) require no auth.
    Task-scoped endpoints require `Authorization: Bearer <session_token>` where
    the token is returned by `POST /tasks`.
    Admin endpoints require `Authorization: Bearer <admin_key>` (static key from
    environment variable).

servers:
  - url: /api/v1
    description: Versioned API root

tags:
  - name: audios
    description: Browse audio assets synced from MSSQL
  - name: tasks
    description: Claim and manage editing locks
  - name: subtitles
    description: Fetch original and submit fixed VTT
  - name: admin
    description: Administrative operations (protected)

# ---------------------------------------------------------------------------
# Reusable components
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: Opaque token returned by POST /tasks. Scoped to a single task.
    adminKey:
      type: http
      scheme: bearer
      description: Static admin key from environment variable ADMIN_API_KEY.

  parameters:
    audioId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Audio ID (sourced from MSSQL, stored in SQLite)

    taskId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Task ID

    submissionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Submission ID

    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0

  schemas:

    Audio:
      type: object
      description: An audio asset registered in the system (metadata from MSSQL).
      required: [id, title, audio_url, status, duration_seconds, synced_at]
      properties:
        id:
          type: string
          format: uuid
          description: Canonical ID from MSSQL.
        title:
          type: string
          example: "Interview with Dr. Chen — 2024-03"
        audio_url:
          type: string
          format: uri
          description: |
            Direct URL to the audio file (from MSSQL). Clients stream this URL
            themselves; the backend never proxies audio.
          example: "https://media.example.com/audios/abc123.mp3"
        duration_seconds:
          type: integer
          example: 1840
        language:
          type: string
          example: "en"
          nullable: true
        status:
          type: string
          enum: [available, in_progress, completed, retired]
          description: |
            - available   — ready to be claimed
            - in_progress — locked by an active task
            - completed   — a submission has been accepted
            - retired     — no longer present in MSSQL; hidden from public listing
        synced_at:
          type: integer
          description: Unix timestamp of the last MSSQL sync for this row.

    AudioList:
      type: object
      required: [data, total, offset, limit]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Audio'
        total:
          type: integer
          description: Total matching audios (for pagination).
        offset:
          type: integer
        limit:
          type: integer

    Task:
      type: object
      required: [task_id, audio_id, status, expires_at]
      properties:
        task_id:
          type: string
          format: uuid
        audio_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, submitted, expired, abandoned]
        expires_at:
          type: integer
          description: Unix timestamp when the lock expires.
        seconds_remaining:
          type: integer
          description: Seconds until expiry (computed on read; absent if not active).

    TaskCreated:
      type: object
      description: Response when a task is successfully created.
      required: [task_id, session_token, audio, expires_at]
      properties:
        task_id:
          type: string
          format: uuid
        session_token:
          type: string
          description: |
            Opaque bearer token the client must include for all task-scoped
            requests. It encodes the task_id and an HMAC signature.
        audio:
          $ref: '#/components/schemas/Audio'
        expires_at:
          type: integer
          description: Unix timestamp when this task expires.
        seconds_remaining:
          type: integer
          example: 7200

    Submission:
      type: object
      required: [submission_id, task_id, audio_id, fixed_s3_key, submitted_at]
      properties:
        submission_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        audio_id:
          type: string
          format: uuid
        fixed_s3_key:
          type: string
          description: S3 key where the fixed VTT was written.
          example: "fixed/abc123/def456.vtt"
        submitted_at:
          type: integer
          description: Unix timestamp of submission.
        byte_size:
          type: integer
          description: Size of the submitted VTT file in bytes.

    SubmissionList:
      type: object
      required: [data, total, offset, limit]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Submission'
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer

    SyncResult:
      type: object
      required: [inserted, updated, retired, duration_ms]
      properties:
        inserted:
          type: integer
          description: New audios added to SQLite from MSSQL.
        updated:
          type: integer
          description: Existing rows whose metadata was refreshed.
        retired:
          type: integer
          description: Rows marked retired because they disappeared from MSSQL.
        duration_ms:
          type: integer
          description: How long the sync took in milliseconds.
        synced_at:
          type: integer
          description: Unix timestamp when the sync completed.

    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: "Audio is not available for claiming"
        code:
          type: string
          description: Machine-readable error code.
          example: "AUDIO_NOT_AVAILABLE"

  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid or expired session token"
            code: "UNAUTHORIZED"

    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Audio not found"
            code: "NOT_FOUND"

    TaskExpired:
      description: The task exists but its lock has expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Task has expired. Claim the audio again to start a new task."
            code: "TASK_EXPIRED"

    Conflict:
      description: The resource is in a state that prevents this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:

  # ---- Audios --------------------------------------------------------------

  /audios:
    get:
      operationId: listAudios
      summary: List audio assets
      description: |
        Returns audios available for the crowd to work on. Defaults to
        `status=available`. Retired audios are never returned unless the caller
        explicitly requests them (admin use only — clients should never need to).
      tags: [audios]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [available, in_progress, completed]
            default: available
          description: Filter by operational status. `retired` is excluded from the enum intentionally.
        - name: language
          in: query
          schema:
            type: string
          description: Filter by language code (e.g. `en`, `he`).
      responses:
        '200':
          description: Paginated list of audios.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioList'

  /audios/{id}:
    get:
      operationId: getAudio
      summary: Get a single audio
      description: Returns full metadata for one audio asset.
      tags: [audios]
      parameters:
        - $ref: '#/components/parameters/audioId'
      responses:
        '200':
          description: Audio details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Audio'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---- Tasks ---------------------------------------------------------------

  /tasks:
    post:
      operationId: createTask
      summary: Claim an audio
      description: |
        Atomically sets the audio to `in_progress` and creates a task that locks
        it for `TASK_TTL_SECONDS` (default 7200). Returns a `session_token` the
        client must include as a bearer token for all subsequent task-scoped
        requests.

        Fails with `409` if the audio is not currently `available`.
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [audio_id]
              properties:
                audio_id:
                  type: string
                  format: uuid
                  description: ID of the audio to claim.
      responses:
        '201':
          description: Task created. Store the session_token — it cannot be retrieved again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /tasks/{id}:
    get:
      operationId: getTask
      summary: Get task status
      description: Returns the current task state and time remaining on the lock.
      tags: [tasks]
      security:
        - sessionToken: []
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: abandonTask
      summary: Abandon a task
      description: |
        Voluntarily releases the lock. Sets task status to `abandoned` and
        returns the audio to `available`. Safe to call more than once.

        The client should call this when the user navigates away or explicitly
        gives up, so the audio becomes available immediately rather than waiting
        for the TTL to expire.
      tags: [tasks]
      security:
        - sessionToken: []
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '204':
          description: Task abandoned successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---- Subtitles -----------------------------------------------------------

  /tasks/{id}/subtitles:
    get:
      operationId: getSubtitles
      summary: Stream original VTT
      description: |
        Fetches the original VTT from S3 and streams it to the client.
        The backend validates the session token before fetching, so S3 is never
        exposed directly for subtitle reads.

        Returns raw VTT bytes — no JSON envelope. The client can pass this URL
        directly to a `<track>` element or fetch it manually.

        Responds with `410 Gone` if the task has expired.
      tags: [subtitles]
      security:
        - sessionToken: []
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Raw VTT content.
          headers:
            Content-Disposition:
              schema:
                type: string
              example: 'inline; filename="audio-abc123.vtt"'
            Cache-Control:
              schema:
                type: string
              example: 'private, no-store'
          content:
            text/vtt:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/TaskExpired'

    post:
      operationId: submitSubtitles
      summary: Submit fixed VTT
      description: |
        Accepts the corrected VTT as the request body. The backend:

        1. Validates the session token and confirms the task is `active`.
        2. Validates that the body is well-formed VTT (starts with `WEBVTT`,
           parses without errors).
        3. Writes the content to S3 at `fixed/<audio_id>/<submission_id>.vtt`.
        4. Inserts a row into `submissions`.
        5. Sets `tasks.status = 'submitted'` and `audios.status = 'completed'`.

        Send the VTT as raw text in the request body with
        `Content-Type: text/vtt`.

        Responds with `410 Gone` if the task has expired.
        Responds with `422 Unprocessable Entity` if the VTT is malformed.
      tags: [subtitles]
      security:
        - sessionToken: []
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          text/vtt:
            schema:
              type: string
              format: binary
              description: |
                The full corrected VTT file content. Must begin with `WEBVTT`.
            example: |
              WEBVTT

              1
              00:00:01.000 --> 00:00:03.500
              Hello, welcome to the show.

              2
              00:00:04.000 --> 00:00:07.200
              Today we are speaking with Dr. Chen.
      responses:
        '200':
          description: Submission accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/TaskExpired'
        '422':
          description: VTT content failed validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "VTT parse error at line 14: missing timestamp separator"
                code: "VTT_INVALID"

  # ---- Admin ---------------------------------------------------------------

  /admin/sync:
    post:
      operationId: triggerSync
      summary: Trigger MSSQL sync
      description: |
        Manually triggers a full sync from MSSQL into SQLite. The same sync
        runs automatically on startup. Use this to pick up newly added audios
        without restarting the service.

        The operation is synchronous — it completes the sync before responding.
        For large datasets, this may take several seconds.
      tags: [admin]
      security:
        - adminKey: []
      responses:
        '200':
          description: Sync completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          description: Could not connect to MSSQL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "MSSQL connection failed: timeout after 10s"
                code: "MSSQL_UNAVAILABLE"

  /admin/audios:
    get:
      operationId: adminListAudios
      summary: List all audios including retired
      description: Same as GET /audios but includes `retired` status and exposes `subtitle_s3_key`.
      tags: [admin]
      security:
        - adminKey: []
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [available, in_progress, completed, retired]
      responses:
        '200':
          description: Paginated audio list including all statuses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioList'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /admin/audios/{id}/reset:
    post:
      operationId: resetAudio
      summary: Reset audio to available
      description: |
        Forces an audio back to `available` and expires any active task on it.
        Use when a submission needs to be redone or a task is stuck.
      tags: [admin]
      security:
        - adminKey: []
      parameters:
        - $ref: '#/components/parameters/audioId'
      responses:
        '200':
          description: Audio reset.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Audio'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /admin/submissions:
    get:
      operationId: listSubmissions
      summary: List all submissions
      description: Returns all accepted submissions, optionally filtered by audio.
      tags: [admin]
      security:
        - adminKey: []
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: audio_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated submissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /admin/submissions/{id}/download:
    get:
      operationId: downloadSubmission
      summary: Download a fixed VTT
      description: |
        Returns a short-lived presigned S3 URL for downloading the fixed VTT
        file associated with a submission.
      tags: [admin]
      security:
        - adminKey: []
      parameters:
        - $ref: '#/components/parameters/submissionId'
      responses:
        '200':
          description: Presigned download URL.
          content:
            application/json:
              schema:
                type: object
                required: [url, expires_in_seconds]
                properties:
                  url:
                    type: string
                    format: uri
                    description: Presigned S3 URL valid for `expires_in_seconds`.
                  expires_in_seconds:
                    type: integer
                    example: 900
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /admin/tasks:
    get:
      operationId: listTasks
      summary: List all tasks
      description: Returns all tasks, useful for monitoring stuck or expired locks.
      tags: [admin]
      security:
        - adminKey: []
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, submitted, expired, abandoned]
      responses:
        '200':
          description: Paginated task list.
          content:
            application/json:
              schema:
                type: object
                required: [data, total, offset, limit]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

  /admin/tasks/sweep:
    post:
      operationId: sweepExpiredTasks
      summary: Manually run expiry sweep
      description: |
        Marks all tasks past their `expires_at` as `expired` and returns their
        locked audios to `available`. The same sweep runs automatically every 15
        minutes. Use this to unblock audios immediately without waiting.
      tags: [admin]
      security:
        - adminKey: []
      responses:
        '200':
          description: Sweep completed.
          content:
            application/json:
              schema:
                type: object
                required: [tasks_expired, audios_released]
                properties:
                  tasks_expired:
                    type: integer
                  audios_released:
                    type: integer
                  swept_at:
                    type: integer
                    description: Unix timestamp.
        '401':
          $ref: '#/components/responses/Unauthorized'